pyproject.toml
[project]
name = "hex"
version = "0.1.0"
description = "AI-enabled virtual spatial proteomics from histopathology for interpretable biomarker discovery in lung cancer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.2.0",
    "captum>=0.7.0",
    "fsspec>=2024.10.0",
    "ftfy>=6.3.1",
    "gitpython>=3.1.43",
    "h5py>=3.12.1",
    "huggingface-hub>=0.26.5",
    "imageio>=2.36.1",
    "joblib>=1.4.2",
    "lifelines>=0.30.0",
    "lightning-utilities>=0.11.9",
    "lxml>=5.3.0",
    "matplotlib>=3.9.3",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy>=2.2.0",
    "opencv-python>=4.10.0.84",
    "openslide-python>=1.4.1",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "protobuf>=5.29.1",
    "pytorch-lightning>=2.2.1",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.5.2",
    "scikit-survival>=0.23.1",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "tensorboardx>=2.6.2.2",
    "timm>=0.9.8",
    "torch-geometric>=2.6.1",
    "tqdm>=4.67.1",
    "transformers>=4.47.0",
    "wandb>=0.19.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
